Move the OWS specific models into their own namespace
