+++ src/Wps/Wps.Client/Models/Ows/ContactInfo.cs
@@ -1,7 +1,7 @@
 ﻿using System.Xml.Serialization;
 using Wps.Client.Utils;
 
-namespace Wps.Client.Models
+namespace Wps.Client.Models.Ows
 {
     /// <summary>
     /// Address of the responsible party.
