+++ src/Wps/Wps.Client/Models/Ows/OperationsMetadata.cs
@@ -1,7 +1,7 @@
 ﻿using System.Xml.Serialization;
 using Wps.Client.Utils;
 
-namespace Wps.Client.Models
+namespace Wps.Client.Models.Ows
 {
     /// <summary>
     /// Metadata about the operations and related abilities specified by this service and implemented by this server.
