+++ src/Wps/Wps.Client/Models/Ows/ServiceIdentification.cs
@@ -1,7 +1,7 @@
 ﻿using System.Xml.Serialization;
 using Wps.Client.Utils;
 
-namespace Wps.Client.Models
+namespace Wps.Client.Models.Ows
 {
     /// <summary>
     /// General metadata for a specific server.
