+++ src/Wps/Wps.Client/Models/Ows/ServiceProvider.cs
@@ -1,7 +1,7 @@
 ﻿using System.Xml.Serialization;
 using Wps.Client.Utils;
 
-namespace Wps.Client.Models
+namespace Wps.Client.Models.Ows
 {
     [XmlRoot("ServiceProvider", Namespace = ModelNamespaces.Ows)]
     public class ServiceProvider
